Refactored WishList to reusable hooks and components
