+++ src/App.tsx
@@ -1,13 +1,14 @@
 import { Routes, Route } from "react-router-dom";
 import Header from "./Header";
 import Search from "./sub-pages/Search/Search";
-import WishList from "./sub-pages/WishList";
+import WishList from "./sub-pages/WishList/WishList";
 import WhatIsVerifind from "./sub-pages/WhatIsVerifind";
 import Login from "./sub-pages/Login";
 import SignUp from "./sub-pages/SignUp";
 import Settings from "./sub-pages/Settings";
 import { SearchProvider } from "./Contexts/SearchContext";
 import { UserProvider } from "./Contexts/UserContext";
+import { WishlistProvider } from "./Contexts/WishListContext";
 
 function App() {
   return (
@@ -15,15 +16,17 @@ function App() {
       <Header />
       <SearchProvider>
         <UserProvider>
-          <Routes>
-            <Route path="/" element={<Search />} />
-            <Route path="/search" element={<Search />} />
-            <Route path="/wish-list" element={<WishList />} />
-            <Route path="/what-is-verifind" element={<WhatIsVerifind />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/signup" element={<SignUp />} />
-            <Route path="/settings" element={<Settings />} />
-          </Routes>
+          <WishlistProvider>
+            <Routes>
+              <Route path="/" element={<Search />} />
+              <Route path="/search" element={<Search />} />
+              <Route path="/wish-list" element={<WishList />} />
+              <Route path="/what-is-verifind" element={<WhatIsVerifind />} />
+              <Route path="/login" element={<Login />} />
+              <Route path="/signup" element={<SignUp />} />
+              <Route path="/settings" element={<Settings />} />
+            </Routes>
+          </WishlistProvider>
         </UserProvider>
       </SearchProvider>
     </>
