+++ src/sub-pages/WishList/wish-list-structures/wishListStructs.ts
@@ -1,0 +1,31 @@
+export type WishlistItem = {
+  //from supabase
+  id: string;
+  product_id: string;
+  product_title: string;
+  product_image: string;
+  target_price: number;
+};
+
+export type EnrichedItem = WishlistItem & {
+  //from serpAPI
+  live_price?: string;
+  rating?: number;
+  reviews?: number;
+  seller?: string;
+  product_url?: string;
+  review_url?: string;
+};
+
+export type OtherWishlistItem = {
+  //from supabase for other people's wishlists
+  product_title: string;
+  product_image: string;
+  target_price: number;
+};
+
+// price history type — from supabase price_history table
+export type PricePoint = {
+  recorded_at: string;
+  price: number;
+};
